Move the remaining kubectl bits to k8s.io/kubectl
