--- pkg/kubectl/cmd/plugin/testdata/kubectl-foo
@@ -1,3 +1,0 @@
-#!/bin/bash
-
-echo "I am plugin foo"
