--- pkg/kubectl/cmd/plugin/testdata/kubectl-version
@@ -1,4 +1,0 @@
-#!/bin/bash
-
-# This plugin is a no-op and is used to test plugins
-# that overshadow existing kubectl commands
